Split worker client into per-family modules behind a barrel
Move the API-family functions out of the 2,293-line client.ts into per-family
modules under client/:
  connection       connect, worker version/instance, state/branches, tool, relay
  firestore-refs   ref/query/sentinel descriptor factories (no RPC)
  firestore-reads  read + aggregate execution + onSnapshot
  firestore-writes writes + writeBatch + runTransaction
  rules            Firestore/RTDB rules deploy + status
  admin-firestore  admin-lens document ops
  rtdb             RTDB modular subset
  studio           event stream, confirm-policy, sandbox snapshot
  auth             per-port auth surface
  storage          worker-backed storage mirror

client.ts is now a 54-line re-export barrel. serve/worker/index.ts and the
pyric-tools/serve/worker subpath export are unchanged. Each family module is
under the 600-line limit (largest is auth at 487). Function bodies are copied
verbatim — pure mechanical move, no behavior change. Export surface verified
byte-identical before/after via a compiler-API probe.
